Refactore filters API by breaking ReportService
Migrate filters api within FiltersService

Ignore test crashes for RxFiltersServiceTest

Refactor ControlsApi to UseCase pattern
