+++ core/src/main/java/com/jaspersoft/android/sdk/service/report/ReportService.java
@@ -25,10 +25,6 @@
 package com.jaspersoft.android.sdk.service.report;
 
 import com.jaspersoft.android.sdk.network.AuthorizedClient;
-import com.jaspersoft.android.sdk.network.entity.control.InputControl;
-import com.jaspersoft.android.sdk.network.entity.control.InputControlState;
-import com.jaspersoft.android.sdk.network.entity.report.ReportParameter;
-import com.jaspersoft.android.sdk.network.entity.report.option.ReportOption;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
 import com.jaspersoft.android.sdk.service.internal.DefaultExceptionMapper;
 import com.jaspersoft.android.sdk.service.internal.Preconditions;
@@ -39,8 +35,6 @@ import com.jaspersoft.android.sdk.service.internal.info.InfoCacheManager;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -53,29 +47,6 @@ public abstract class ReportService {
                                         @Nullable ReportExecutionOptions execOptions) throws ServiceException;
 
     @NotNull
-    public abstract List<InputControl> listControls(@NotNull String reportUri) throws ServiceException;
-
-    @NotNull
-    public abstract List<InputControlState> listControlsValues(@NotNull String reportUri,
-                                                               @NotNull List<ReportParameter> parameters) throws ServiceException;
-
-    @NotNull
-    public abstract Set<ReportOption> listReportOptions(@NotNull String reportUri) throws ServiceException;
-
-    @NotNull
-    public abstract ReportOption createReportOption(@NotNull String reportUri,
-                                                    @NotNull String optionLabel,
-                                                    @NotNull List<ReportParameter> parameters,
-                                                    boolean overwrite) throws ServiceException;
-
-    public abstract void updateReportOption(@NotNull String reportUri,
-                                            @NotNull String optionId,
-                                            @NotNull List<ReportParameter> parameters) throws ServiceException;
-
-    public abstract void deleteReportOption(@NotNull String reportUri,
-                                            @NotNull String optionId) throws ServiceException;
-
-    @NotNull
     public static ReportService newService(@NotNull AuthorizedClient client) {
         Preconditions.checkNotNull(client, "Client should not be null");
 
@@ -86,8 +57,6 @@ public abstract class ReportService {
         ReportServiceFactory reportServiceFactory = new ReportServiceFactory(cacheManager,
                 client.reportExecutionApi(),
                 client.reportExportApi(),
-                client.inputControlApi(),
-                client.reportOptionsApi(),
                 reportMapper,
                 client.getBaseUrl(),
                 TimeUnit.SECONDS.toMillis(1)
