+++ core/src/main/java/com/jaspersoft/android/sdk/service/report/ReportService5_6Plus.java
@@ -31,11 +31,9 @@ package com.jaspersoft.android.sdk.service.report;
 final class ReportService5_6Plus extends AbstractReportService {
     protected ReportService5_6Plus(ExportExecutionApi exportExecutionApi,
                                    ReportExecutionApi reportExecutionApi,
-                                   ReportOptionsUseCase reportOptionsUseCase,
-                                   ControlsApi controlsApi,
                                    ExportFactory exportFactory,
                                    long delay) {
-        super(exportExecutionApi, reportExecutionApi, reportOptionsUseCase, controlsApi, exportFactory, delay);
+        super(exportExecutionApi, reportExecutionApi, exportFactory, delay);
     }
 
     @Override
