+++ rx/src/main/java/com/jaspersoft/android/sdk/service/rx/report/RxReportService.java
@@ -25,10 +25,6 @@
 package com.jaspersoft.android.sdk.service.rx.report;
 
 import com.jaspersoft.android.sdk.network.AuthorizedClient;
-import com.jaspersoft.android.sdk.network.entity.control.InputControl;
-import com.jaspersoft.android.sdk.network.entity.control.InputControlState;
-import com.jaspersoft.android.sdk.network.entity.report.ReportParameter;
-import com.jaspersoft.android.sdk.network.entity.report.option.ReportOption;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
 import com.jaspersoft.android.sdk.service.internal.Preconditions;
 import com.jaspersoft.android.sdk.service.report.ReportExecution;
@@ -39,9 +35,6 @@ import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
 import rx.Observable;
 import rx.functions.Func0;
-
-import java.util.List;
-import java.util.Set;
 
 /**
  * @author Tom Koptel
@@ -69,118 +62,6 @@ public class RxReportService {
                 } catch (ServiceException e) {
                     return Observable.error(e);
                 }
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<List<InputControl>> listControls(@NotNull final String reportUri) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-
-        return Observable.defer(new Func0<Observable<List<InputControl>>>() {
-            @Override
-            public Observable<List<InputControl>> call() {
-                try {
-                    List<InputControl> inputControls = mSyncDelegate.listControls(reportUri);
-                    return Observable.just(inputControls);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<List<InputControlState>> listControlsValues(@NotNull final String reportUri,
-                                                                  @NotNull final List<ReportParameter> parameters) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-        Preconditions.checkNotNull(parameters, "Parameters should not be null");
-
-        return Observable.defer(new Func0<Observable<List<InputControlState>>>() {
-            @Override
-            public Observable<List<InputControlState>> call() {
-                try {
-                    List<InputControlState> inputControlStates = mSyncDelegate.listControlsValues(reportUri, parameters);
-                    return Observable.just(inputControlStates);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<Set<ReportOption>> listReportOptions(@NotNull final String reportUri) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-
-        return Observable.defer(new Func0<Observable<Set<ReportOption>>>() {
-            @Override
-            public Observable<Set<ReportOption>> call() {
-                try {
-                    Set<ReportOption> reportOptions = mSyncDelegate.listReportOptions(reportUri);
-                    return Observable.just(reportOptions);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<ReportOption> createReportOption(@NotNull final String reportUri,
-                                                       @NotNull final String optionLabel,
-                                                       @NotNull final List<ReportParameter> parameters,
-                                                       final boolean overwrite) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-        Preconditions.checkNotNull(optionLabel, "Option label should not be null");
-        Preconditions.checkNotNull(parameters, "Parameters should not be null");
-
-        return Observable.defer(new Func0<Observable<ReportOption>>() {
-            @Override
-            public Observable<ReportOption> call() {
-                try {
-                    ReportOption reportOption = mSyncDelegate.createReportOption(reportUri, optionLabel, parameters, overwrite);
-                    return Observable.just(reportOption);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<Void> updateReportOption(@NotNull final String reportUri, @NotNull final String optionId, @NotNull final List<ReportParameter> parameters) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-        Preconditions.checkNotNull(optionId, "Option id should not be null");
-        Preconditions.checkNotNull(parameters, "Parameters should not be null");
-
-        return Observable.defer(new Func0<Observable<Void>>() {
-            @Override
-            public Observable<Void> call() {
-                try {
-                    mSyncDelegate.updateReportOption(reportUri, optionId, parameters);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-                return Observable.just(null);
-            }
-        });
-    }
-
-    @NotNull
-    public Observable<Void> deleteReportOption(@NotNull final String reportUri, @NotNull final String optionId) {
-        Preconditions.checkNotNull(reportUri, "Report uri should not be null");
-        Preconditions.checkNotNull(optionId, "Option id should not be null");
-
-        return Observable.defer(new Func0<Observable<Void>>() {
-            @Override
-            public Observable<Void> call() {
-                try {
-                    mSyncDelegate.deleteReportOption(reportUri, optionId);
-                } catch (ServiceException e) {
-                    return Observable.error(e);
-                }
-                return Observable.just(null);
             }
         });
     }
