+++ rx/src/test/java/com/jaspersoft/android/sdk/service/rx/report/RxReportServiceTest.java
@@ -25,10 +25,6 @@
 package com.jaspersoft.android.sdk.service.rx.report;
 
 import com.jaspersoft.android.sdk.network.AuthorizedClient;
-import com.jaspersoft.android.sdk.network.entity.control.InputControl;
-import com.jaspersoft.android.sdk.network.entity.control.InputControlState;
-import com.jaspersoft.android.sdk.network.entity.report.ReportParameter;
-import com.jaspersoft.android.sdk.network.entity.report.option.ReportOption;
 import com.jaspersoft.android.sdk.service.exception.ServiceException;
 import com.jaspersoft.android.sdk.service.report.ReportExecution;
 import com.jaspersoft.android.sdk.service.report.ReportExecutionOptions;
@@ -40,26 +36,18 @@ import org.junit.rules.ExpectedException;
 import org.mockito.Mock;
 import rx.observers.TestSubscriber;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
-
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.rules.ExpectedException.none;
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyBoolean;
-import static org.mockito.Matchers.anyListOf;
 import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 public class RxReportServiceTest {
     private static final String REPORT_URI = "my/uri";
     private static final ReportExecutionOptions OPTIONS = ReportExecutionOptions.builder().build();
-    private static final List<ReportParameter> PARAMS = Collections.emptyList();
-    private static final String OPTION_LABEL = "label";
-    private static final String OPTION_ID = OPTION_LABEL;
 
     @Mock
     ReportService mSyncDelegate;
@@ -67,8 +55,6 @@ public class RxReportServiceTest {
     ReportExecution mReportExecution;
     @Mock
     ServiceException mServiceException;
-
-    private ReportOption fakeReportOption = new ReportOption();
 
     @Mock
     AuthorizedClient mAuthorizedClient;
@@ -127,169 +113,6 @@ public class RxReportServiceTest {
         verify(mSyncDelegate).run(REPORT_URI, OPTIONS);
     }
 
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_list_controls() throws Exception {
-        when(mSyncDelegate.listControls(anyString())).thenThrow(mServiceException);
-
-        TestSubscriber<List<InputControl>> test = TestSubscriber.create();
-        rxReportService.listControls(REPORT_URI).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).listControls(REPORT_URI);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_list_controls() throws Exception {
-        when(mSyncDelegate.listControls(anyString())).thenReturn(Collections.<InputControl>emptyList());
-
-        TestSubscriber<List<InputControl>> test = TestSubscriber.create();
-        rxReportService.listControls(REPORT_URI).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).listControls(REPORT_URI);
-    }
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_list_controls_values() throws Exception {
-        when(mSyncDelegate.listControlsValues(anyString(), anyListOf(ReportParameter.class))).thenThrow(mServiceException);
-
-        TestSubscriber<List<InputControlState>> test = TestSubscriber.create();
-        rxReportService.listControlsValues(REPORT_URI, PARAMS).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).listControlsValues(REPORT_URI, PARAMS);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_list_controls_values() throws Exception {
-        when(mSyncDelegate.listControlsValues(anyString(), anyListOf(ReportParameter.class)))
-                .thenReturn(Collections.<InputControlState>emptyList());
-
-        TestSubscriber<List<InputControlState>> test = TestSubscriber.create();
-        rxReportService.listControlsValues(REPORT_URI, PARAMS).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).listControlsValues(REPORT_URI, PARAMS);
-    }
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_list_report_options() throws Exception {
-        when(mSyncDelegate.listReportOptions(anyString())).thenThrow(mServiceException);
-
-        TestSubscriber<Set<ReportOption>> test = TestSubscriber.create();
-        rxReportService.listReportOptions(REPORT_URI).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).listReportOptions(REPORT_URI);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_list_report_options() throws Exception {
-        when(mSyncDelegate.listReportOptions(anyString()))
-                .thenReturn(Collections.<ReportOption>emptySet());
-
-        TestSubscriber<Set<ReportOption>> test = TestSubscriber.create();
-        rxReportService.listReportOptions(REPORT_URI).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).listReportOptions(REPORT_URI);
-    }
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_create_report_option() throws Exception {
-        when(mSyncDelegate.createReportOption(anyString(), anyString(), anyListOf(ReportParameter.class), anyBoolean()))
-                .thenThrow(mServiceException);
-
-        TestSubscriber<ReportOption> test = TestSubscriber.create();
-        rxReportService.createReportOption(REPORT_URI, OPTION_LABEL, PARAMS, true).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).createReportOption(REPORT_URI, OPTION_LABEL, PARAMS, true);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_create_report_option() throws Exception {
-        when(mSyncDelegate.createReportOption(anyString(), anyString(), anyListOf(ReportParameter.class), anyBoolean()))
-                .thenReturn(fakeReportOption);
-
-        TestSubscriber<ReportOption> test = TestSubscriber.create();
-        rxReportService.createReportOption(REPORT_URI, OPTION_LABEL, PARAMS, true).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).createReportOption(REPORT_URI, OPTION_LABEL, PARAMS, true);
-    }
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_update_report_option() throws Exception {
-        doThrow(mServiceException).when(mSyncDelegate).updateReportOption(anyString(),
-                anyString(), anyListOf(ReportParameter.class));
-
-        TestSubscriber<Void> test = TestSubscriber.create();
-        rxReportService.updateReportOption(REPORT_URI, OPTION_ID, PARAMS).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).updateReportOption(REPORT_URI, OPTION_ID, PARAMS);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_update_report_option() throws Exception {
-        TestSubscriber<Void> test = TestSubscriber.create();
-        rxReportService.updateReportOption(REPORT_URI, OPTION_ID, PARAMS).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).updateReportOption(REPORT_URI, OPTION_ID, PARAMS);
-    }
-
-    @Test
-    public void should_delegate_service_exception_to_subscription_on_delete_report_option() throws Exception {
-        doThrow(mServiceException).when(mSyncDelegate).deleteReportOption(anyString(), anyString());
-
-        TestSubscriber<Void> test = TestSubscriber.create();
-        rxReportService.deleteReportOption(REPORT_URI, OPTION_ID).subscribe(test);
-
-        test.assertError(mServiceException);
-        test.assertNotCompleted();
-
-        verify(mSyncDelegate).deleteReportOption(REPORT_URI, OPTION_ID);
-    }
-
-    @Test
-    public void should_execute_delegate_as_observable_on_delete_report_option() throws Exception {
-        TestSubscriber<Void> test = TestSubscriber.create();
-        rxReportService.deleteReportOption(REPORT_URI, OPTION_ID).subscribe(test);
-
-        test.assertCompleted();
-        test.assertNoErrors();
-        test.assertValueCount(1);
-
-        verify(mSyncDelegate).deleteReportOption(REPORT_URI, OPTION_ID);
-    }
 
     @Test
     public void should_provide_impl_with_factory_method() throws Exception {
